@@ -15,78 +15,10 @@ use Doctrine\Persistence\ManagerRegistry;
  * @method OrderProduct[]    findAll()
  * @method OrderProduct[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-class OrderProductRepository extends ServiceEntityRepository
+class OrderProductRepository extends BaseRepository
 {
-    /**
-     * @throws Exception
-     */
     public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, OrderProduct::class);
-        $this->_em->getConnection()->beginTransaction();
     }
-
-    /**
-     * @throws Exception
-     */
-    public function add(OrderProduct $entity, bool $flush = false): void
-    {
-        $this->getEntityManager()->persist($entity);
-
-        if ($flush) {
-            $this->getEntityManager()->flush();
-            $this->_em->getConnection()->commit();
-        }
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function remove(OrderProduct $entity, bool $flush = false): void
-    {
-        $this->getEntityManager()->remove($entity);
-
-        if ($flush) {
-            $this->getEntityManager()->flush();
-            $this->_em->getConnection()->commit();
-        }
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function update(OrderProduct $entity, bool $flush = false): void
-    {
-        $this->getEntityManager()->persist($entity);
-
-        if ($flush) {
-            $this->getEntityManager()->flush();
-            $this->_em->getConnection()->commit();
-        }
-    }
-
-//    /**
-//     * @return OrderProduct[] Returns an array of OrderProduct objects
-//     */
-//    public function findByExampleField($value): array
-//    {
-//        return $this->createQueryBuilder('o')
-//            ->andWhere('o.exampleField = :val')
-//            ->setParameter('val', $value)
-//            ->orderBy('o.id', 'ASC')
-//            ->setMaxResults(10)
-//            ->getQuery()
-//            ->getResult()
-//        ;
-//    }
-
-//    public function findOneBySomeField($value): ?OrderProduct
-//    {
-//        return $this->createQueryBuilder('o')
-//            ->andWhere('o.exampleField = :val')
-//            ->setParameter('val', $value)
-//            ->getQuery()
-//            ->getOneOrNullResult()
-//        ;
-//    }
 }
